MCP/requirements.txt
# requirements.txt
google-generativeai==0.3.2
mysql-connector-python==8.1.0